Remove Airflow and write to S3
